Add rate limiting to UDP sockets.
This is necessary to slow mavtables down when other components can't
keep up with the packet rate and need mavtables to buffer packets.  The
max_bitrate setting lets the end user restrict the output UDP bitrate to
a given bits per second, avoiding overflowing the operating system
buffer when the other components are slow.